+++ js/form.js
@@ -1,5 +1,6 @@
 import {typeProperty} from './generation.js';
 import {postData} from './api.js';
+import {successPost, failPost} from './popup.js';
 
 const form = document.querySelector('.ad-form');
 const pristine = new Pristine(form, {
@@ -23,7 +24,6 @@ function validateAmount () {
   //const capacitySelectValue = String(capacitySelect.value);
   return roomCapacityMap[roomNumberSelect.value].includes(capacitySelect.value);
 }
-
 
 function getAmountErrorMessage () {
   let message = 'Количество гостей не может превышать количество комнат';
@@ -86,67 +86,6 @@ pristine.addValidator(
   1,
   true
 );
-
-const body = document.querySelector('body');
-// Находим фрагмент с содержимым темплейта
-const successPopup = document.querySelector('#success')
-  .content
-  .querySelector('.success');
-
-// Находим фрагмент с содержимым темплейта
-const failPopup = document.querySelector('#error')
-  .content
-  .querySelector('.error');
-const closeButton = failPopup.querySelector('.error__button');
-
-// Проверка клавиши esc
-const isEscapeKey = (evt) => evt.key === 'Escape';
-
-// Если нажали esc, то закрываем форму
-const onPopupEscKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    oncloseFailPopup();
-    oncloseSuccessPopup();
-  }
-};
-
-// Функция закрытия успешного сообщения
-function oncloseSuccessPopup () {
-  body.removeChild(successPopup);
-  document.removeEventListener('keydown', onPopupEscKeydown);
-  successPopup.removeEventListener('click', oncloseSuccessPopup);
-}
-
-// Функция закрытия неуспешного сообщения
-function oncloseFailPopup () {
-  body.removeChild(failPopup);
-  document.removeEventListener('keydown', onPopupEscKeydown);
-  failPopup.removeEventListener('click', oncloseFailPopup);
-  closeButton.removeEventListener('click', oncloseFailPopup);
-}
-
-
-//успешная отправка
-const successPost = () => {
-  //Находим расположение successTemplate и отрисовываем шаблон там
-  body.appendChild(successPopup);
-  // Добавляем обработчики на закрытие сообщения
-  document.addEventListener('keydown', onPopupEscKeydown);
-  successPopup.addEventListener('click', oncloseSuccessPopup);
-  form.reset();
-};
-
-
-//Неуспешная отправка
-const failPost = () => {
-  //Находим расположение successTemplate и отрисовываем шаблон там
-  body.appendChild(failPopup);
-  // Добавляем обработчики на закрытие сообщения
-  document.addEventListener('keydown', onPopupEscKeydown);
-  failPopup.addEventListener('click', oncloseFailPopup);
-  closeButton.addEventListener('click', oncloseFailPopup);
-};
 
 //Отправка формы
 form.addEventListener('submit', (evt) => {
